Implement QPlatformTheme functionality
This allows same functionality on all desktop Platforms:
 * macOS
 * Linux
 * Windows

On macOS the `QIcon::setThemeName` setup was needed, but with a
_"folder-hack.svg"_ instead of _"folder.svg"_. This was needed so that
Qt would ask for a `QPlatformTheme::createIconEngine` which would return
a custom SVGIconOffOnEngine that would render both `QIcon::Off` and
`QIcon::On` states.

On Windows the `QPlatformTheme::standardPixmap` was needed so that the
`QPlatformTheme::DirIcon` and `QPlatformTheme::DirOpenIcon` would be
rendered.

Lastly on Linux the `QIcon::setThemeName` was required not be set.
Otherwise there would be no possibility to get the "folder-open.svg" set
as `QIcon::On` for `folder.svg`.